Move parse_image_path
